Rename logger name
